+++ lib/pscom4extoll/pscom_extoll.c
@@ -103,7 +103,7 @@ int pscom_extoll_rma2_do_read(pscom_poll_reader_t *reader)
 static
 void pscom_extoll_rma2_do_write(pscom_con_t *con)
 {
-	unsigned int len;
+	size_t len;
 	struct iovec iov[2];
 	pscom_req_t *req;
 
@@ -113,7 +113,7 @@ void pscom_extoll_rma2_do_write(pscom_con_t *con)
 		psex_con_info_t *ci = con->arch.extoll.ci;
 		len = iov[0].iov_len + iov[1].iov_len;
 
-		int rlen = psex_sendv(ci, iov, len);
+		ssize_t rlen = psex_sendv(ci, iov, len);
 
 		if (rlen >= 0) {
 			pscom_write_done(con, req, rlen);
